Update /ignore messaging
Add message if you're not ignoring anything.

Add more messages to be suppressed by Ignores.DrawOutput

Already shown by prompt.
    "Place or break two blocks to determine the edges."

No news is good news.
    "Created a new " + scope + " custom block"...
    "Duplicated the {0} custom block "...
    "Copied the {0} custom block "...

Any error or warning will be shown, but if the message is expected it
is suppressed, for example, a missing block for block remove command is
not a problem.

Add blockdef ignores flag

Rename to BlockdefChanges